feat: Copy from item page template
