+++ erpnext/stock/doctype/item_page_template/test_item_page_template.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2020, Frappe Technologies Pvt. Ltd. and Contributors
+# See license.txt
+from __future__ import unicode_literals
+
+# import frappe
+import unittest
+
+class TestItemPageTemplate(unittest.TestCase):
+	pass
